requirements.txt
transformers
torch
fastapi
uvicorn
gradio
torch
transformers
grpcio
grpcio-tools
pydantic
requests
edge-tts
uuid
asyncio
